+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Configure all tasks that extend task classes instead of just those created by the plugin 
 - Make JbrResolver prefer Gradle javaToolchains by `JetBrains s.r.o`, if available. Only otherwise start fetching and running a new one.
 - Support for Kotlin Coroutines debugging
+- Detect and warn if project adds an explicit dependency on Kotlin Coroutines library
 
 ### Changed
 - Disabled caching for `BuildPluginTask`